Proj_v23/requirements.txt
Flask
Werkzeug
openai
frontend
tools
youtube_transcript_api
langdetect
translate
langchain
googletrans==4.0.0-rc1
bcrypt
Flask-SQLAlchemy
datetime
Flask-Mail
Jinja2
chardet
fitz